+++ ScrapMe/ScrapMe.cs
@@ -7,7 +7,9 @@ using RoR2;
 namespace ScrapMe
 {
 
-    // This attribute is required, and lists metadata for your plugin.
+    /// <summary>
+    /// Main plugin class. Named so you can invoke a using directive on the namespace.
+    /// </summary>
     [BepInPlugin(PluginGUID, PluginName, PluginVersion)]
     [BepInDependency("com.rune580.riskofoptions", BepInDependency.DependencyFlags.SoftDependency)]
     [BepInDependency("com.Gorakh.ItemQualities", BepInDependency.DependencyFlags.SoftDependency)]
@@ -15,12 +17,13 @@ namespace ScrapMe
     {
         /// singleton
         public static ScrapMe plugin;
-        // why are these not project constants
+        
+    #pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
         public const string PluginGUID = PluginAuthor + "." + PluginName;
         public const string PluginAuthor = "not_score";
         public const string PluginName = "ScrapMe";
         public const string PluginVersion = "0.3.0";
-        
+    #pragma warning restore CS1591 // Missing XML comment for publicly visible type or member        
         
         
         internal ConfigManager configManager;
@@ -40,79 +43,26 @@ namespace ScrapMe
             RoR2Application.onLoad += OnLoad;
         }
 
-        public void OnLoad()
+        private void OnLoad()
         {
-            configManager.Load(); // in here now. lets every mod load first
+            PresetBans(); // load dev bans first to leverage the quality compat update
+            configManager.Load(); // in here now, to let every character and item load first
             RiskOfOptionsCompat.InitConfigMenu();
-            PresetBans();
         }
 
         /// <summary>
         /// Gets the developer-set bans for a given character.
+        /// After changing these bans, make sure to call <see cref="QualityCompat.SetQualityVariantBans(BodyIndex)"/>
         /// </summary>
         /// <param name="bodyIndex">Name of the body's prefab.</param>
-        /// <returns>HashSet for performing set operations on</returns>
-        public HashSet<ItemIndex> GetDevBans(BodyIndex bodyIndex)
-        {
-            if (!devItemBans.ContainsKey(bodyIndex))
-            {
-                devItemBans[bodyIndex] = new();
-                mappedBodies.Add(bodyIndex);
-            }
-
-            return devItemBans[bodyIndex];
-        }
-
-        /// <summary>
-        /// Set the bans of a particular character.
-        /// </summary>
-        /// <param name="bodyIndex">Name of the body's prefab.</param>
-        /// <param name="bans">Collection of PickupDef.internalName's for items to be banned.</param>
-        public void SetDevBans(BodyIndex bodyIndex, IEnumerable<ItemIndex> bans)
-        {
-            devItemBans[bodyIndex] = new(bans);
-            mappedBodies.Add(bodyIndex);
-        }
-
-        public void SetDevBans(string bodyName, IEnumerable<string> bans)
-        {
-            var bodyIndex = BodyCatalog.FindBodyIndex(bodyName);
-            if (bodyIndex == BodyIndex.None)
-            {
-                Log.Warning($"Couldn't add bans for body {bodyName}; wasn't able to resolve name in catalog");
-                return;
-            }
-
-            HashSet<ItemIndex> items = new();
-            foreach (var ban in bans)
-            {
-                var itemIndex = ItemCatalog.FindItemIndex(ban);
-                if (itemIndex == ItemIndex.None)
-                {
-                    Log.Info($"Couldn't find definition for item {ban}, skipping");
-                }
-                else
-                {
-                    items.Add(itemIndex);
-                }
-            }
-            SetDevBans(bodyIndex, items);
-        }
-
+        /// <returns>HashSet for performing set operations on, or null if the index provided was None</returns>
+        public HashSet<ItemIndex> GetDevBans(BodyIndex bodyIndex) => bodyIndex == BodyIndex.None ? null : bans.dev[bodyIndex];
+        
         internal void PresetBans()
         {
-            SetDevBans("RobBelmontBody",["BarrierOnCooldown","JumpBoost","JumpDamageStrike"]);
-            SetDevBans("RobRavagerBody",["JumpBoost","JumpDamageStrike"]);
+            BanRecords.presets.ForEach(preset => preset.Resolve());
         }
-        
-        internal readonly Dictionary<BodyIndex, HashSet<ItemIndex>> devItemBans = new();
 
-        internal readonly Dictionary<BodyIndex, HashSet<ItemIndex>> userItemBans = new();
-
-        internal readonly Dictionary<BodyIndex, HashSet<ItemIndex>> userItemUnbans = new();
-
-        internal readonly Dictionary<BodyIndex, HashSet<ItemIndex>> qualityBans = new();
-        
-        internal readonly HashSet<BodyIndex> mappedBodies = [];
+        internal readonly BanRecords bans = new();
     }
 }