+++ ScrapMe/Utils.cs
@@ -1,33 +1,35 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using JetBrains.Annotations;
 using RoR2;
 
 namespace ScrapMe;
-
+/// <summary>
+/// General utility functions for the plugin.
+/// </summary>
 public static class Utils
 {
-    [NotNull]
-    public static HashSet<ItemIndex> GetBans(BodyIndex bodyIndex, bool withQuality = true)
-    {
-        if (bodyIndex == BodyIndex.None) return [];
-        var bans = ScrapMe.plugin.devItemBans[bodyIndex]
-            .Union(ScrapMe.plugin.userItemBans[bodyIndex]);
-        if (withQuality)
-        {
-            bans = bans.Union(ScrapMe.plugin.qualityBans[bodyIndex]);
-        }
-        return bans.Except(ScrapMe.plugin.userItemUnbans[bodyIndex]).ToHashSet();
-    }
 
+    /// <summary>
+    /// Gets the name of a prefab from an item.
+    /// </summary>
+    /// <param name="itemName">Name of the item.</param>
+    /// <returns>Name of the prefab.</returns>
+    [Obsolete]
     public static string GetPrefabNameFromItem(string itemName)
     {
         return itemName.Substring(itemName.IndexOf("Index.") + 6);
     }
 
+    /// <summary>
+    /// Internal caching dictionary for item tiers. DO NOT TOUCH.
+    /// </summary>
     private static Dictionary<ItemTier, ItemIndex> scrapLookup = [];
     
+    /// <summary>
+    /// Gets the correct scrap item for the relevant tier. Should only be called if necessary for logging reasons.
+    /// </summary>
+    /// <param name="tier"></param>
+    /// <returns></returns>
     public static ItemIndex GetScrapForTier(ItemTier tier)
     {
         if (!scrapLookup.ContainsKey(tier))
@@ -65,29 +67,47 @@ public static class Utils
     }
 
     
-    
+    /// <summary>
+    /// Gets the correct replacement item for a given item, applying all relevant compatibility layers.
+    /// </summary>
+    /// <param name="item"></param>
+    /// <returns></returns>
     public static ItemIndex GetReplacementItem(ItemIndex item)
     {
         var scrap = GetScrapForTier(ItemCatalog.GetItemDef(item).tier);
         return QualityCompat.CarryQualityToNewItem(item, scrap);
     }
     
+    /// <summary>
+    /// Gets the name of a character prefab from a clone of that prefab.
+    /// </summary>
+    /// <param name="cloneName"></param>
+    /// <returns></returns>
     public static string GetPrefabNameFromClone(string cloneName)
     {
         return cloneName.Substring(0, cloneName.LastIndexOf("(Clone)"));
     }
     
+    /// <summary>
+    /// Intended to sanitize character prefab names, for the inevitable day someone reports an issue with Cho'Gath.
+    /// </summary>
+    /// <param name="prefabName"></param>
+    /// <returns></returns>
     public static string SanitizePrefabName(string prefabName)
     {
-        // TODO
+        // TODO upon bug report
         return prefabName;
     }
 
+    /// <summary>
+    /// Scrubs the player's inventory of any illegal items. DOES NOT CHECK IF THERE ARE VOID CORRUPTIONS.
+    /// </summary>
+    /// <param name="body"></param>
     public static void CheckInventory(CharacterBody body)
     {
         if (body == null) return;
         
-        var bans = GetBans(body.bodyIndex);
+        var bans = ScrapMe.plugin.bans.All(body.bodyIndex);
         foreach (var itemBan in bans)
         {
             var itemCount = body.inventory.GetItemCountPermanent(itemBan);